create a RootNav component - this serves as the main component of the App. - uses a nav drawer - currently just shows dummy text
